fix: remove unused workshop page
